Reexport pokemons page from index
